refactor(request_helpers): Moved BoundedStream to stream module (#1583).
