+++ docs/api/request_and_response_wsgi.rst
@@ -40,6 +40,9 @@ Request
 .. autoclass:: falcon.Forwarded
     :members:
 
+.. autoclass:: falcon.stream.BoundedStream
+    :members:
+
 Response
 --------
