+++ falcon/media/multipart.py
@@ -18,8 +18,8 @@ import cgi
 import re
 from urllib.parse import unquote_to_bytes
 
+import falcon.stream
 from falcon import errors
-from falcon import request_helpers
 from falcon.media.base import BaseHandler
 from falcon.util import BufferedReader, DelimiterError
 from falcon.util import deprecated_args, misc
@@ -251,7 +251,7 @@ class MultipartForm:
         # This approach makes testing both the Cythonized and pure-Python
         #   streams easier within the same test/benchmark suite.
         if not hasattr(stream, 'read_until'):
-            if isinstance(stream, request_helpers.BoundedStream):
+            if isinstance(stream, falcon.stream.BoundedStream):
                 stream = BufferedReader(stream.stream.read, content_length)
             else:
                 stream = BufferedReader(stream.read, content_length)
